support for hadoop properties in faunus.properties. added combiners to degree and edge distribution (based on IntWritable now -- may need to move to LongWritable). Stubbed support for pre-graph load filtering.

+++ src/main/java/com/thinkaurelius/faunus/mapreduce/operators/DegreeDistribution.java
@@ -5,7 +5,6 @@ import com.thinkaurelius.faunus.util.Tokens;
 import com.tinkerpop.blueprints.Direction;
 import com.tinkerpop.blueprints.Edge;
 import org.apache.hadoop.io.IntWritable;
-import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.NullWritable;
 import org.apache.hadoop.mapreduce.Mapper;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -24,7 +23,7 @@ public class DegreeDistribution {
         EDGES_COUNTED
     }
 
-    public static class Map extends Mapper<NullWritable, FaunusVertex, LongWritable, IntWritable> {
+    public static class Map extends Mapper<NullWritable, FaunusVertex, IntWritable, IntWritable> {
 
         private Direction direction;
         private String[] labels;
@@ -36,27 +35,27 @@ public class DegreeDistribution {
         }
 
         @Override
-        public void map(final NullWritable key, final FaunusVertex value, final Mapper<NullWritable, FaunusVertex, LongWritable, IntWritable>.Context context) throws IOException, InterruptedException {
-            long degree = 0l;
+        public void map(final NullWritable key, final FaunusVertex value, final Mapper<NullWritable, FaunusVertex, IntWritable, IntWritable>.Context context) throws IOException, InterruptedException {
+            int degree = 0;
             for (final Edge edge : value.getEdges(this.direction, this.labels)) {
                 degree++;
             }
 
             context.getCounter(Counters.EDGES_COUNTED).increment(degree);
-            context.write(new LongWritable(degree), new IntWritable(1));
+            context.write(new IntWritable(degree), new IntWritable(1));
 
         }
 
     }
 
-    public static class Reduce extends Reducer<LongWritable, IntWritable, LongWritable, LongWritable> {
+    public static class Reduce extends Reducer<IntWritable, IntWritable, IntWritable, IntWritable> {
         @Override
-        public void reduce(final LongWritable key, final Iterable<IntWritable> values, final Reducer<LongWritable, IntWritable, LongWritable, LongWritable>.Context context) throws IOException, InterruptedException {
-            long totalDegree = 0l;
+        public void reduce(final IntWritable key, final Iterable<IntWritable> values, final Reducer<IntWritable, IntWritable, IntWritable, IntWritable>.Context context) throws IOException, InterruptedException {
+            int totalDegree = 0;
             for (final IntWritable token : values) {
                 totalDegree++;
             }
-            context.write(key, new LongWritable(totalDegree));
+            context.write(key, new IntWritable(totalDegree));
         }
     }
 }
